Converted a bunch of commands to a dialog
Pressing the quick action key opens the Micahcraft dialog, which lets you do all the old location and leaderboard commands. The location command have been removed, but the leaderboard ones remain. The horse commands have been left alone, as well as AFK, Settings and the creative ones. Settings can be opened with either the command or in the pause menu